+++ manifest.config.ts
@@ -29,4 +29,5 @@ export default defineManifest({
   side_panel: {
     default_path: 'src/sidepanel/index.html',
   },
+  options_page: 'src/options/index.html',
 })